Favor hard links over file copies where suitable.
After extracting and/or flattening the custom tracks, all files are
guaranteed to live in a temporary directory in a single file system,
which enables the creation of hard links.

In tests on Linux, when a retail ISO image (PAL) is patched with the
default arguments [and an arbitrary set of distinct custom tracks],
compilation time has dropped from `24 seconds` to `17 seconds` (average
of 30 runs).

On Windows, compilation time has dropped from `26 seconds to
`19 seconds` (best of 5 runs).