Changed memory address for temporary flag that prevents button spam.
Previously, the memory address of the Gecko register #10 (grA) was used.
However, this was causing problems when the game was played in modded
consoles; it seems critical memory was being overridden, causing the
game to crash/stall when the race started, even though course page
selection appeared to work.

The new memory address is region-specific, and it should be safe to
modify, as this memory address actually points to padding data that is
embedded in the `.dol` file.

Thanks to @BlackTSQ for reporting and helping diagnosing the issue
thoroughly, and testing the fix.