Added support for the 6 extra audio tracks that may be used by the stock courses in the input ISO file.
The MKDD Extender was not yet aware of the 6 extra audio tracks that a
non-vanilla input ISO file may be utilizing. This is the case with ISO
files that have been treated by the MKDD Track Patcher before being
feeded to the MKDD Extender.

When the audio track file indexes were gathered, the order could break
if audio files in the ISO file's file list were in a non-asciibetical
order.

Thanks to Katerina Angela for reporting the issue.