Add option to disregard custom audio tracks in favor of the audio track of the retail course defined by the `replaces` field in the INI file.
This is the **Use Replacee Audio Track** (`--use-replacee-audio-track`)
option.

Also, the fallback audio track, used when custom audio tracks are not
provided, is now the audio track of the retail course defined in the
`replaces` field, instad of the audio track of Luigi Circuit (previous
behavior). If the `replaces` field is not provided, Luigi Circuit's
audio track is still the fallback audio track.

Closes #8.