Cast page number to `int` before applying modulo operation in the C code.
When the page count (`PAGE_COUNT`) was set to for example `3` (instead
of `4`, or even `2`), the assembly code produced by devkitPPC did not
seem to function as expected: the user could no longer decrement the
page number, only increment it.

The current hypothesis is that `char` variables do not get along with
the `%` operator. At this is point it is unclear whether this is
expected behavior, or caused by [unknown] UB around these lines, or
whether it is a bug in the compiler.