Auto-correct BTI images with unexpected wrap S/T values.
Preview and logo images in some custom tracks may show unexpected values
in the 5th and 6th bytes in the BTI header (usually `0x01` and `0x01`
respectively), which will be manifested in image corruption when
Nintendont is used.

The values will be zeroed (i.e. set to `0x00` and `0x00`).