Remove potential hard link before attempting to fix up BTI header.
This was a regression in 7fd6e10ef3fe8e928a78fdce9e39ee6c72362e40.

When a BTI image had unexpected values in the wrap S/T fields, the
associated stock image in the stock course would be inadvertently
replaced with the image in the custom track after the MKDD Extender
applies the fix to the BTI header.

For performance reasons, hard links to the stock files are used when
placeholders are needed. If the hard link is not broken, any attempt to
modify one of the references would result in the original image being
replaced too.