Add logic for verifying that required code patches are present and enabled.
This is only error handling; no code patch exists yet.

When a custom track requires code patches (comma-separated list of code
patch names in the `code_patches` field in the `trackinfo.ini` file),
the MKDD Extender will verify that the code patch exists (i.e. it is
supported), and that it has been enabled. If the code patch is not
enabled, an error is raised. A *dangerous* option is offered to
circumvent the error (for expert users that know what they are doing).