Avoid defining a static array in memory and use pre-existing padding.
Instead of defining `player_item_rolls` as a static array, which
occupies extra space in the DOL file, and requires its own symbol,
pre-existing *padding* in the DOL file is now repurposed to store the
data required for the **Type-specific Item Boxes** code patch.

Number of instructions is reduced from `132` (528 bytes) to `55` (220
bytes).