Add option for specifying the initial course page.
This is the **Initial Page Number** option, which defaults to `1` (the
stock courses in the input ISO file).

The option allows users to skip the first page in the ISO file. It can
be helpful while developing a new custom track to avoid the extra button
press required to select the page with the work-in-progress track.