Invalidate instruction blocks after changing minimap orientation values.
Minimap coordinates and orientations are hardcoded in code. In order to
change these values, code instructions need to be modified when the user
changes the course page. However, part of these instructions can be
cached (e.g. by Dolphin's JIT compiler), and subsequent modifications
are not picked up.

This was specifically a problem for the minimap orientations, which are
hardcoded in `li` instructions, and updating the operand on the fly
when the course page changes did not make a difference, as the
instruction, and its operands, had been cached.

In order to invalidate the instruction block, `icbi` instructions
targeting the address of the `li` instruction are now inserted.

Other instructions, such as `dcbf`, or `sync` and `isync`, are inserted
too, as they were used in the example seen at https://www.ibm.com/docs/en/aix/7.1?topic=set-icbi-instruction-cache-block-invalidate-instruction.
On Dolphin, the `icbi` instruction alone is sufficient to invalidate the
cache, but it is thought that, on real hardware, the other instructions
may be needed, in particular, the `dcbf` instruction.