Improve padding in numbered cup name images.
Previously, images with a width of 362 pixels were generated, leaving
the downscaling [to 256 pixels wide] to the game paiting engine. This
approach was not only less spacially efficient, but also enforced
downscaling even in cup name images that already had enough marging
(notably, the cup name images in the NTSC-J version).

The new approach keeps the canvas size to 256 pixels wide, downscaling
the cup name only if extra margin is needed for overlaying the page
number / page count.