Distinguish between "Luigi2" and "Mini2" to avoid dropping the suffix in the latter.
For Luigi Circuit 100cc+, because the course name has a "2" suffix in
it, which is not present in the internal directory name, the suffix is
removed. The slot for Nintendo GameCube (Mini2) was inadvertently
affected by the removal of the suffix; it took quite a few hours to spot
this subtle problem; custom battle stages in this slot would simply not
work!