Embed page number and page count in the preview image of the first battle stage in every course page.
In the **SELECT COURSE** and **SELECT CUP** screens, the current page
number, and the total page count is embedded in the cup names, as it is
possible to refresh/reload those images on page change. However, no such
element exists in the **SELECT STAGE** screen.

As an alternative, the current page number / page count is now embedded
in the preview images, which get updates properly on page change.