Set media source in media player after slots have been connected.
Since Qt 6.5.0, the media is read (or at least cued) synchronously as
part of the `QMediaPlayer.setSource()` call; the
`QMediaPlayer.seekableChanged` and `QMediaPlayer.durationChanged`
signals are emitted then. If the slots are not connected before the
source is set, the events are missed. These events are currently used to
enable the media player's slider.

This behavior change in Qt 6.5.0 was reported in
https://bugreports.qt.io/browse/QTBUG-119446.