Handle node entries of type `0x9500` in RARC module.
Until now, only node entries of type `0x1100` (file type) or `0x0200`
(directory type) were handled. It has then been noticed that another
entry type can be encountered: `0x9500`.

This type, which is not documented in the reference specifications (see
module's docstrings) seemingly relates to the YAZ0/SZS formats. The
solution for now is to treat it in the same way as `0x1100`.