Check that page items in settings file is not empty before attempting to operate on the list.
In settings files that have gone bad (usually as a result of other
issues that will be addressed in follow-up revisions), the
`page_item_combined_values` list could remain valid, but empty, causing
an exception when attempting to apply `max()` on an empty list:

```
Traceback (most recent call last):
  File "mkdd_extender.py", line 3214, in <module>
  File "mkdd_extender.py", line 3194, in main
  File "gui.py", line 3913, in run
  File "gui.py", line 2032, in __init__
  File "gui.py", line 2186, in _restore_settings
ValueError: max() arg is an empty sequence
[13977] Failed to execute script 'mkdd_extender' due to unhandled exception!
```

This prevented the process from starting. In a follow-up commit, safety
measures will be put in place to guarantee exceptions raised while
restoring settings are caught.