Add the **Use Alternative Buttons** option.
When enabled, `Y` and `X` will be set as the buttons for switching
course pages in the game.

In LAN mode, these buttons (unlike the `D-pad Up` and `D-pad Down`
buttons that are used by default) will be transmitted through the
network, meaning that course page switching will be synchronized between
the host and the rest of the consoles.