Rework based on the `baa` module how the BSFT file inside `GCKart.baa` is updated.
Previously, the embedded BSFT file in `GCKart.baa` was updated by
_appending_ another copy of it, and overriding its data offset in the
header entry. This was not ideal, but it was functional.

Now, `GCKart.baa` is unpacked first, then its `.bsft` is rewritten, and
the BAA file is repacked.

Additionally, the standalone `GCKart.bsft` file (next to `GCKart.baa` in
the `AudioRes` directory) is now updated too. Although the game does not
access this file, it is known that some third-party modding tools assume
that this file matches the embedded file in `GCKart.baa`.