Add **Remove Movie Trailer** option.
When enabled, the `files/Movies` directory is removed from the resulting
ISO file, and `main.dol` is modified to skip the game screens that would
attempt to load `play1.thp` or `play1_50.thp`.

Note that code changes were not totally necessary, but it prevents the
Lakitu/cloud loading screen that would otherwise continue to appear even
if the movie trailer does not get loaded or played.