Only create progress dialogs when the operation is taking long.
Previously, the dialog would be created regardless, only to be shown
after 100 ms if the background operation was still ongoing.

However, this was causing issues when an error was produced early in the
worker thread (e.g. a pre-extended input ISO file was detected): the
exception was caught and presented to the user correctly, but then the
rogue progress dialog that was never shown would appear, and couldn't be
closed. Who would have thought that the mix of Qt's event loops, the
worker thread, and the re-raised exception caught in one thread but
raised in another could cause problems!

Now the actual progress dialog widget is only created when it is known
that it will be shown immediately, after 100 ms.