Prevent spurious page switches in LAN mode when alternative buttons are used.
When the **Use Alternative Buttons** option is used (a6d9cb10), an alternative
memory region that holds the _synchronized_ state of the game buttons is
used in LAN mode. This 1-byte region is used to determine whether the
main player has pressed either the `Y` / `X` buttons.

However, it was noticed that, every few minutes, while sitting in the
**SELECT MODE** screen in LAN mode, course pages could switch
automatically without the intervention of the main player. Some times
the course page could switch multiple times within a few seconds.
Although all connected consoles would attempt to switch pages at the
same time, this was not guaranteed, and they would desync after some
rouge course page switches.

After monitoring the aforementioned memory region, it was eventually
observed that, when course pages switched automatically, the memory
region was randomly flipping between `00000000` and `00001101`. The
third and fourth bits match the `Y` and `X` buttons' mask values.

It is not known what `00001101` represents, and why it appears to be
broadcast to all connected consoles in the sesion. Keepalive messages?
Heartbeats? Something else?

In order to mitigate the effects of the spurious data, the previous
bitwise operation to determine whether either of the buttons is pressed
has been made stricter; the full byte is now compared against each mask
value.