Reuse "SELECT COURSE" menu title image in LAN mode.
The "SELECT MODE" screen is shared between multiplayer and LAN mode; it
could not be modified without altering also the use in multiplayer.

In order to provide the menu title image in LAN mode with the controls
image (something that was skipped in the initial work), the "SELECT
COURSE" image will be used, which already includes the controls icon.
The `li` instruction that holds the hardcoded index will be tweaked
(incremented by `1`) to point the other image.

Note that, in a previous commit (e00273ffed4b607e), a different approach
was taken, where the "VS.? CO-OP?" image was repurposed, transferring
ownserhip to the LAN mode, which could then be modified freely. However,
users argued that the "SELECT COURSE" text was even more suitable, which
happens to simplify the logic. This other approach was disregarded; the
commit was reverted to keep the details in the repository history.