Remove fragile implementation of the mechanism for playing the course page change sound.
This implementation used the `JAISeMgr::startSound()` function to play
the sound. However, the arguments for the function, which are pointers
to heap memory, were hardcoded in the Python module: the pointers had to
be adjusted with an offset based on the OS Arena relocation.

It was fragile, and crashes had been reported when the Extender was used
on non-stock ISOs.

A different implementation will be introduced in a follow-up commit.