Reintroduce course page change sound.
A previous implementation was available until 5698bc2e; it was fragile,
and has been dropped in favor of this implementation.

This new implementation uses `GameAudio::Main::startSystemSe()`, a
a method in what appears to be a singleton class. Its first argument
is the pointer to the static, singleton class; the second argument is
the game ID of the sound to be played.