Firstly, removed all "== true" and "== false" conditionals for booleans. Using ! for false I find can be easy to miss at times, but it's also what I ordinarily do anyway.
Set a bunch of variables in functions to be 'consts'. There are quite a lot of variables in my code that
are changed, and then returned. On a related note, I'm absolutely terrible at one-liners.

Made sure to use 'f' suffix for 'float'. Wish it was a warning. Unity tells me, and I appreciate it for that.