Add support for code patch "Bouncy Terrain Type". (#25)
Adds "Bouncy Terrain Type" to list of code patches.

This patch adds custom terrain type, 0xB0, which will add force to the upwards and forwards direction of a Kart to achieve a bounce. It achieves this effect by hijacking "KartStrat::DoSpeedCrl" to stop it from clamping the force added, allowing for movements that were never possible in the vanilla game. 

Tested on all retail versions via Dolphin. Patch does nothing if applied to Debug version.

This patch has a github repository dedicated to documenting the terrain type it adds. It can be found here: 
https://lance-o.github.io/extended_terrain_types/bouncy_terrain_type/BOUNCY_TERRAIN_TYPE.html
https://lance-o.github.io/extended_terrain_types/