Add automatic code patching mode.
A new **Code Patching Mode** option has been introduced, which defaults
to `Automatic`: code patches that are needed by at least one of the
configured custom courses will be automatically enabled.

The options for the code patches intended for custom courses are now
disabled if **Code Patching Mode** is not set to `Manual`.