Disable section highlighting in courses list.
Since there is a single **Custom Courses** column, there is little
benefit with always highlighting (bold style) the column header.

Notice that, as per the removed inline comment, this was achieved in the
past in some hacky way, but it did not appear to be working anymore in
Qt 6.5; perhaps the hack never really worked.