Defer paint update of the course list after the filter has been applied to all rows.
`showRow()` and `hideRow()` appear to be triggering a paint update
reactively on every call. When the visibility is changed in many rows at
once (especially when showing rows after they had been hidden), the
application could stall for a noticeable while. For instance, with ~200
courses in the list, and the filter previously set to "A", the
application would freeze for ~300 ms once the filter was cleared.

The solution is to disable paint updates via `setUpdatesEnabled()` until
all rows have been updated, and then trigger a manual update if
required.