Add support for 0 extra course pages.
Until now, it was required that at least one extra course page was
configured, to a total of two course pages.

To better support code patches and part of the option set, this
limitation has been lifted, allowing users to build an ISO with no extra
course pages added (i.e. the initial 16 race tracks and 6 battle stages
remain the only courses in the input ISO, but manually-enabled code
patches and options are still applied).