Support for cheat codes that write to memory areas not defined in the DOL header.
Not all cheat codes write to memory sections that are well defined in
the DOL header. This is fine when a code handler is available, as the
code handler will apply the cheat code after the game's OS has been
initialized.

However, in MKDD Patcher, since no code handler will be installed, all
cheat codes need to be baked into a proper section. In the absence of
such section, MKDD Patcher will try to find the section that precedes
the address where the data needs to be written and resize it.

The following mod can be used for testing this functionality:

[`EverythingUnlocked.zip`](https://github.com/user-attachments/files/18629128/EverythingUnlocked.zip)

The cheat code in the mod (which writes to unmanaged memory areas) will
unlock all cups, courses, characters, karts, and game modes.