Stricter logic to locate minimap image for previewing purposes.
Previously, the minimap image in custom courses was selected by
searching for any file whose filename ends with `_map.bti`; that would
be the image that is shown in the **Minimap** section in the preview
pane.

However, it was overlooked that, in Mushroom City (and thus in custom
courses that use it as a base), there is a second image that matches the
pattern: `patapata_course_map.bti`. Depending on the file ordering that
the filesystem in the OS provides, the _other_ image could be
encountered before the actual minimap image; this was the case on
Windows systems.

The logic is a bit stricter now: only images that have exactly one `_`
character in the filename are considered.