Replace assertion with warning in RARC module.
It has been reported that there are RARC archives that can be found in
the wild that feature directory nodes whose entry index in the header is
set wrongly (`0x0000` instead of the expected `0xFFFF`).

The original `assert` has now been replaced with a warning message in
order to support these dubious files.