Check whether overlay widget in `DropWidget` exists before deletion attempt.
Starting with Qt 6.7, `dragLeaveEvent()` may be invoked even if the
event in the associated `dragStartEvent()` call was not accepted. It is
not known at this point whether this is a regression in Qt or a
deliberate behavior change.

With the adoption of Qt 6.8 in 98dbac6f, a benign exception could be
seen in the terminal/console when the user attempted to drag a course
from the courses list onto one of the pages tables, as the internal
overlay widget only exists in `dragLeaveEvent()` if the associated event
in the `dragStartEvent()` call was accepted:

```
Traceback (most recent call last):
  File "/w/mkdd-extender/gui.py", line 692, in dragLeaveEvent
    self._overlay_widget.deleteLater()
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
AttributeError: 'NoneType' object has no attribute 'deleteLater'
```