Add scroll area to shelf items that occupy a large portion of the available screen geometry.
In some systems, depending on the display's resolution and on the
system's font size, the tables in the shelf item tool tips could exceed
the available vertical height, causing some artifacts in the HTML
(normally when many course pages had been configured).

A scroll area is now created in these particular cases; in normal cases
a reqular label is still used.