LAN Choose Character and Kart v1.1
There is a counter on the top right of the menu displaying how many
karts have finished progression out of all karts e.g. 3/8

The host can force the character selection to be exited by holding B.
B button graphic is also displayed as a warning in top right of screen.

Subsequent races will skip the entry screen and go straight to the
character/kart selection screen.

There is also improvement for the RNG seed and other slight code
changes.