Support for pre and post-patching user scripts.
Two new options have been added to the application arguments:

- **Pre-patching Script**
- **Post-patching Script**

They permit the execution of arbitrary Python code before and after
MKDD Extender applies the modifications to the input ISO image. Users
are given an opportunity to apply their more specific requirements that
are not general enough to be included in MKDD Extender with opt-in
preferences.