Reintroduce the GeckoLoader tool as a Git submodule.
In the origins of MKDD Extender, GeckoLoader was used for injecting
the custom logic that made possible switching between course pages.
Eventually that mechanism was superseded by the current GC-C-Kit-based
approach.

Now, in order to introduce a new feature for injecting rich Gecko codes
directly in MKDD Extender, the GeckoLoader tool is being added back.