Logic to migrate cheat codes-related settings from the previous version.
The support for the new, fully-featured Gecko cheat codes (see previous
commit) introduced a breaking change in the user settings: cheat codes
configured/enabled in a previous release would be lost when launching
the future release.

To reduce the nuisance that that brings, this temporary commit will
upgrade the user settings so that the values set in a previous release
are respected in the future release.